More changes to the class Player
Must update class Hand to be able to discard externally. This is needed
for splitting. Maybe it would be more convenient to create a split
method in hand that checks that one has a pair (valid splitting), leaves
a card in the hand and returns a new hand containing the other card.